Reorganize tool_discovery
